Refactor entities, add activity logic and meals randomizer

+++ src/main/java/com/solutionscrafted/nutriplanner/dto/RecipeDto.java
@@ -7,7 +7,7 @@ public record RecipeDto(
         Long id,
         String name,
         String instructions,
-        Double totalCalories,
+        Integer totalCalories,
         String tags
 ) {
 }
